% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_draw.R
\name{draw_key}
\alias{draw_key}
\alias{draw_key_boxFull}
\alias{draw_key_boxLeft}
\alias{draw_key_boxRight}
\alias{draw_key_boxSplit}
\alias{draw_key_violinFull}
\alias{draw_key_violinSplit}
\alias{draw_key_violinLeft}
\alias{draw_key_violinRight}
\alias{draw_number_circle}
\title{Key glyphs for legends}
\usage{
draw_key_boxFull(data, params, size)

draw_key_boxLeft(data, params, size)

draw_key_boxRight(data, params, size)

draw_key_boxSplit(data, params, size)

draw_key_violinFull(data, params, size)

draw_key_violinSplit(data, params, size)

draw_key_violinLeft(data, params, size)

draw_key_violinRight(data, params, size)

draw_number_circle(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob.
}
\description{
Each geom has an associated function that draws the key when
the geom needs to be displayed in a legend. These functions are
called  draw_key_*(), where * stands for the name of the
respective key glyph.
see examples below.
}
