% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jjviomap.R
\docType{data}
\name{exp.long}
\alias{exp.long}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{data.frame} with 71226 rows and 3 columns.
}
\usage{
exp.long
}
\description{
This is a test data for this package
test data describtion
}
\author{
Junjun Lao
}
\keyword{datasets}
