% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arch2.R
\name{geom_arch2}
\alias{geom_arch2}
\title{Draw an Arch2 shape}
\usage{
geom_arch2(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  arrow = NULL,
  arrow.fill = "black",
  lineend = "butt",
  ndivisions = 200,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}}.}

\item{data}{A dataset to use for the plot.}

\item{stat}{The statistical transformation to use on the data.}

\item{position}{The position adjustment to use for overlapping points.}

\item{...}{description}

\item{arrow}{The type of arrowhead to use for the ends of the curve.}

\item{arrow.fill}{The fill color for the arrowheads.}

\item{lineend}{The type of line end to use for the curve.}

\item{ndivisions}{The number of divisions used to create the curve.}

\item{na.rm}{If \code{TRUE}, remove missing values from the data before plotting.}

\item{show.legend}{Logical. Should this layer be included in the legends?}

\item{inherit.aes}{If \code{TRUE}, inherit aesthetics from parent plot object.}
}
\value{
A layer that can be added to a ggplot object using the + operator.
}
\description{
Draw an Arch2 shape
}
\author{
Jun Zhang

An Arch2 geom is a curved line with arrowheads at both ends.
}
