% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cross_axis.R
\name{geom_cross_axis}
\alias{geom_cross_axis}
\title{Cross Axis Geom}
\usage{
geom_cross_axis(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  bk.len = 0.02,
  label.size = 3,
  label.shift = 0.015,
  lineend = "butt",
  linejoin = "round",
  arrow.fill = "black",
  colour = "black",
  arrow = NULL,
  fontface = "plain",
  draw.minor.bk = FALSE,
  xline.pos = NULL,
  yline.pos = NULL,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{na.rm}{description}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{bk.len}{The length of the tick marks extending from the axis, as a
proportion of the range of the data. Default is 0.02.}

\item{label.size}{The font size of the labels for the tick marks. Default is 3.}

\item{label.shift}{The distance between the tick mark and its label. Default
is 0.015.}

\item{lineend}{The line ending style for the cross-axis lines. Possible values
are "round", "butt", or "square". Default is "butt".}

\item{linejoin}{The line joining style for the cross-axis lines. Possible
values are "round", "mitre", or "bevel". Default is "round".}

\item{arrow.fill}{The fill color for any arrowheads on the cross-axis lines.
Default is "black".}

\item{colour}{The color of the cross-axis lines. Default is "black".}

\item{arrow}{A logical value indicating whether or not arrowheads should be
drawn at the end of the cross-axis lines. Default is NULL.}

\item{fontface}{The font face for the tick mark labels. Possible values are
"plain", "italic", "bold", or "bold.italic". Default is "plain".}

\item{draw.minor.bk}{A logical value indicating whether or not to draw minor
tick marks. Default is FALSE.}

\item{xline.pos}{A vector of values representing the positions of the vertical
cross-axis lines.}

\item{yline.pos}{A vector of values representing the positions of the
horizontal cross-axis lines.}

\item{...}{description}
}
\description{
Cross Axis Geom
}
\examples{
library(ggplot2)

# Create a simple plot with cross-axis lines
p <- ggplot(iris, aes(Sepal.Width, Petal.Length)) +
  geom_point()
p + geom_cross_axis()

}
\author{
Jun Zhang

The geom_cross_axis() function allows you to add cross-axis lines and labels
to a ggplot2 plot.
}
