% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jjboxplot.R
\name{geom_jjboxplot}
\alias{geom_jjboxplot}
\title{geom_jjboxplot}
\usage{
geom_jjboxplot(
  mapping = NULL,
  data = NULL,
  stat = "jjboxplot",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  outlier.show = TRUE,
  outlier.colour = "grey50",
  outlier.color = "grey50",
  outlier.fill = "grey50",
  outlier.shape = 19,
  outlier.size = 1.5,
  outlier.stroke = 0.5,
  outlier.alpha = 1,
  shift = 0,
  split.shift = 0,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{show.legend}{show.legend}

\item{na.rm}{na.rm}

\item{inherit.aes}{inherit.aes}

\item{outlier.show}{outlier.show}

\item{outlier.colour}{outlier.colour}

\item{outlier.color}{outlier.color}

\item{outlier.fill}{outlier.fill}

\item{outlier.shape}{outlier.shape}

\item{outlier.size}{outlier.size}

\item{outlier.stroke}{outlier.stroke}

\item{outlier.alpha}{outlier.alpha}

\item{shift}{shift}

\item{split.shift}{split.shift}

\item{...}{...}
}
\description{
geom_jjboxplot
}
\author{
Junjun Lao
}
