% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jjpie.R
\name{geom_jjpie}
\alias{geom_jjpie}
\title{geom_jjpie}
\usage{
geom_jjpie(
  mapping = NULL,
  data = NULL,
  stat = "jjpie",
  position = "identity",
  ...,
  value_max = 1,
  add.circle = TRUE,
  circle.color = "black",
  circle.fill = "white",
  pie.theta = NULL,
  add.rect = FALSE,
  rect.width = 1,
  rect.height = 1,
  rect.col = "black",
  rect.fill = "white",
  circle.radius = 0,
  hollow.fill = "white",
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{...}{...}

\item{value_max}{value_max}

\item{add.circle}{add.circle}

\item{circle.color}{circle.color}

\item{circle.fill}{circle.fill}

\item{pie.theta}{pie.theta}

\item{add.rect}{add.rect}

\item{rect.width}{rect.width}

\item{rect.height}{rect.height}

\item{rect.col}{rect.col}

\item{rect.fill}{rect.fill}

\item{circle.radius}{circle.radius}

\item{hollow.fill}{hollow.fill}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}
}
\description{
geom_jjpie
}
\author{
Junjun Lao
}
