% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jjPointPie.R
\name{geom_jjPointPie}
\alias{geom_jjPointPie}
\title{geom_jjPointPie}
\usage{
geom_jjPointPie(
  mapping = NULL,
  data = NULL,
  stat = "jjPointPie",
  position = "identity",
  ...,
  explode.dist = 0.1,
  add.circle = FALSE,
  circle.rev = FALSE,
  circle.colour = "black",
  circle.fill = "white",
  circle.radius = 0.05,
  add.text = FALSE,
  text.all = TRUE,
  text.dist = 0.1,
  text.size = 5,
  ratio.digit = 2,
  text.hjust = 0.5,
  text.vjust = 0.5,
  seq.length = 500,
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{...}{...}

\item{explode.dist}{explode.dist}

\item{add.circle}{add.circle}

\item{circle.rev}{circle.rev}

\item{circle.colour}{circle.colour}

\item{circle.fill}{circle.fill}

\item{circle.radius}{circle.radius}

\item{add.text}{add.text}

\item{text.all}{text.all}

\item{text.dist}{text.dist}

\item{text.size}{text.size}

\item{ratio.digit}{ratio.digit}

\item{text.hjust}{text.hjust}

\item{text.vjust}{text.vjust}

\item{seq.length}{seq.length}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}
}
\description{
geom_jjPointPie
}
\author{
Junjun Lao
}
