% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jjtile.R
\name{geom_jjtile}
\alias{geom_jjtile}
\title{geom_jjtile}
\usage{
geom_jjtile(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  linejoin = "mitre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{...}{...}

\item{linejoin}{linejoin}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}
}
\description{
geom_jjtile
}
\author{
Junjun Lao
}
