% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jjtriangle.R
\name{geom_jjtriangle}
\alias{geom_jjtriangle}
\title{geom_jjtriangle}
\usage{
geom_jjtriangle(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  type = "ul",
  rect = FALSE,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{type}{type}

\item{rect}{rect}

\item{...}{...}
}
\description{
geom_jjtriangle
}
\author{
Junjun Lao
}
