% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jjviolin.R
\name{geom_jjviolin}
\alias{geom_jjviolin}
\title{geom_jjviolin}
\usage{
geom_jjviolin(
  mapping = NULL,
  data = NULL,
  stat = "jjviolin",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  shift = 0,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{show.legend}{show.legend}

\item{na.rm}{na.rm}

\item{inherit.aes}{inherit.aes}

\item{shift}{shift}

\item{...}{...}
}
\description{
geom_jjviolin
}
\author{
Junjun Lao
}
