% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jjviomap.R
\name{geom_jjviomap}
\alias{geom_jjviomap}
\title{geom_jjviomap}
\usage{
geom_jjviomap(
  mapping = NULL,
  data = NULL,
  stat = "jjviomap",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{show.legend}{show.legend}

\item{na.rm}{na.rm}

\item{inherit.aes}{inherit.aes}

\item{...}{...}
}
\description{
geom_jjviomap
}
\author{
Junjun Lao
}
