% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_funcs.R
\name{grid.colorkey}
\alias{grid.colorkey}
\title{Create a Color Key Grid}
\usage{
grid.colorkey(
  x = NULL,
  color = NULL,
  color.n = 100,
  ticks.side = c("left", "right", "top", "bottom"),
  pos = c("h", "v")
)
}
\arguments{
\item{x}{Numeric vector specifying the range for the x-axis.}

\item{color}{Vector of color values to be used for the color key.}

\item{color.n}{Number of colors to generate.}

\item{ticks.side}{Side of the color key to display tick marks
("left", "right", "top", or "bottom").}

\item{pos}{Position of the color key ("h" for horizontal, "v" for vertical).}
}
\description{
This function creates a color key grid using the grid graphics system.
}
