% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_funcs.R
\name{grid.yaxis2}
\alias{grid.yaxis2}
\title{grid.yaxis2 function}
\usage{
grid.yaxis2(
  at = NULL,
  breaks = 5,
  labels = NULL,
  tick.len = 0.5,
  label.space = 0.25,
  side = c("left", "right"),
  rot = 0
)
}
\arguments{
\item{at}{numeric vector specifying the locations of tick marks on the axis.}

\item{breaks}{integer specifying the number of breaks on the axis.}

\item{labels}{character vector providing custom labels for the tick marks.}

\item{tick.len}{numeric specifying the length of the tick marks.}

\item{label.space}{numeric specifying the space between the tick marks and labels.}

\item{side}{character specifying the side to place the axis ("left" or "right").}

\item{rot}{numeric specifying the rotation angle for the tick mark labels.}
}
\description{
grid.yaxis2 function
}
\author{
Jun Zhang

This function creates a secondary y-axis in a graphical grid.
}
