% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_jjboxplot.R
\name{stat_jjboxplot}
\alias{stat_jjboxplot}
\title{stat_jjboxplot}
\usage{
stat_jjboxplot(
  mapping = NULL,
  data = NULL,
  geom = "jjboxplot",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{geom}{geom}

\item{position}{position}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{...}{...}
}
\description{
stat_jjboxplot
}
\author{
Junjun Lao
}
