% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_jjviomap.R
\name{stat_jjviomap}
\alias{stat_jjviomap}
\title{stat_jjviomap}
\usage{
stat_jjviomap(
  mapping = NULL,
  data = NULL,
  geom = "jjviomap",
  position = "identity",
  show.legend = NA,
  trim = TRUE,
  bw = "nrd0",
  adjust = 1,
  kernel = "gaussian",
  n = 512,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{geom}{geom}

\item{position}{position}

\item{show.legend}{show.legend}

\item{trim}{trim}

\item{bw}{bw}

\item{adjust}{adjust}

\item{kernel}{kernel}

\item{n}{n}

\item{inherit.aes}{inherit.aes}

\item{...}{...}
}
\description{
stat_jjviomap
}
\author{
Junjun Lao
}
