% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_jjboxplot.R, R/geom_jjpie.R,
%   R/geom_jjpie2.R, R/geom_jjtriangle.R, R/geom_jjviolin.R, R/stat_jjboxplot.R,
%   R/stat_jjpie.R, R/stat_jjviolin.R
\docType{data}
\name{GeomJjboxplot}
\alias{GeomJjboxplot}
\alias{GeomJjpie}
\alias{geom_jjpie2}
\alias{GeomJjtriangle}
\alias{GeomJjviolin}
\alias{StatJjboxplot}
\alias{StatJjpie}
\alias{StatJjviolin}
\title{geom_jjpie2}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{...}{...}

\item{value_max}{value_max}

\item{add.circle}{add.circle}

\item{circle.color}{circle.color}

\item{circle.fill}{circle.fill}

\item{pie.theta}{pie.theta}

\item{add.rect}{add.rect}

\item{rect.width}{rect.width}

\item{rect.height}{rect.height}

\item{rect.col}{rect.col}

\item{rect.fill}{rect.fill}

\item{circle.radius}{circle.radius}

\item{hollow.fill}{hollow.fill}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}
}
\description{
geom_jjpie2
}
\author{
Junjun Lao
}
\keyword{datasets}
